requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
Flask>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
